requirements.txt
pytube
git+https://github.com/openai/whisper.git
transformers
coqui-tts# Add Python dependencies here 
